demo/pyproject.toml
[project]
name = "sec-agent-demo"
version = "0.1.0"
description = "Behaviour Programming vs Tool Calling on SEC 10-K filings"
requires-python = ">=3.12"
dependencies = [
    "opensymbolicai-core",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "beautifulsoup4>=4.12",
    "pydantic>=2.0",
    "rich>=13.0",
    "numpy>=1.26",
    "httpx>=0.27",
    "lancedb>=0.20",
    "pyarrow>=18.0",
    "textual>=0.40",
]

[tool.uv]
dev-dependencies = []
